pyproject.toml
[project]
name = "qwen3-tts-tokenizer-12hz-trainer"
version = "0.0.1"
description = "Qwen3 TTS Tokenizer 12Hz Trainer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "transformers==4.57.3",
  "accelerate==1.12.0",
  "gradio",
  "librosa",
  "torchaudio",
  "soundfile",
  "sox",
  "onnxruntime",
  "einops",
  "ipykernel",
  "ipywidgets",
  "flash-attn; sys_platform == 'linux'",
  "pyarrow>=23.0.0",
  "webdataset>=1.0.2",
  "wandb>=0.24.1",
  "tensorboard>=2.20.0",
  "peft>=0.18.1",
  "datasets>=4.5.0",
  "black>=26.1.0",
  "matplotlib>=3.10.8",
  "qwen-tts>=0.1.1",
  "loguru>=0.7.3",
  "resampy>=0.4.3",
  "torchcodec>=0.10.0",
]

# third-party whl: https://github.com/Dao-AILab/flash-attention/issues/2299
[tool.uv.sources]
flash-attn = { url = "https://github.com/lesj0610/flash-attention/releases/download/v2.8.3-cu12-torch2.10-cp312/flash_attn-2.8.3%2Bcu12torch2.10cxx11abiTRUE-cp312-cp312-linux_x86_64.whl" }

